Add files via upload
This is a clustering code that initially applies a Pearsons correlation analysis in 3D. In this analysis, a set of 2D gaussian functions of increasing size are compared continuously to the exterior of intensity image stacks describing t-cells. Using the information collected from this analysis, a custom clustering method applies watershedding-baed approach to identify clusters. The Examine Clusters code is a rapid means to evaluate how well the clustering performed on your data.